requirements.txt
# ============================================================
# SENTRY_TRADER — Dependencies
# ============================================================

# Binance API & WebSocket
python-binance==1.0.19
websockets>=12.0

# Async HTTP
aiohttp>=3.9

# Data Processing
pandas>=2.2
numpy>=1.26
ta>=0.11.0           # Technical indicators (RSI, EMA, ATR)

# Database
sqlalchemy>=2.0
aiosqlite>=0.20

# Configuration
python-dotenv>=1.0

# AI/ML (Phase 2)
scikit-learn>=1.4
joblib>=1.3

# Logging
loguru>=0.7

# Telegram (Phase 5)
python-telegram-bot>=21.0

# Utilities
pytz>=2024.1
